pyproject.toml
[tool.poetry]
name = "nycu-sp-2023-fall"
version = "0.1.0"
description = ""
authors = ["LolaInTa Chi <lolainta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pwntools = "^4.11.0"
pycryptodome = "^3.19.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
